Added a couple of (bash) scripts, expanded Shell class and successfully running Package class passed as file via Builder

+++ scripts/build_toolchain_two.sh
@@ -20,13 +20,14 @@ binutils_stage_two() {
         "$SRCDIR"/configure \
             --prefix="$source_packager_builder_environment" \
             --build=$("$SRCDIR"/config.guess) \
+            --host="$source_packager_target" \
+            --disable-nls \
             --enable-shared \
             --disable-werror \
             --enable-64-bit-bfd;
         make -j$source_packager_number_of_jobs && \
         make -j$source_packager_number_of_jobs install;
     cd -;
-    rm -rf "$BLDDIR";
 }
 
 gcc_stage_two() {
@@ -36,67 +37,38 @@ gcc_stage_two() {
     BLDDIR="/tmp/build-$NAME";
     SRCDIR="$source_packager_source_dir/$NAME"
 
-    # if [ -d "$SRCDIR" ]; then
-    #     rm -rf "$SRCDIR";
-    # fi;
+    if [ -d "$SRCDIR" ]; then
+        rm -rf "$SRCDIR";
+    fi;
 
-    # if [ -d "$BLDDIR" ]; then
-    #     rm -rf "$BLDDIR";
-    # fi;
+    if [ -d "$BLDDIR" ]; then
+        rm -rf "$BLDDIR";
+    fi;
 
-    # from_tz_file "$FILEURL" "$NAME";
-    # mkdir "$BLDDIR" -pv;
+    from_tz_file "$FILEURL" "$NAME";
+    mkdir "$BLDDIR" -pv;
 
     cd "$SRCDIR";
-        #sh contrib/download_prerequisites;
-        #sed -e '/m64=/s/lib64/lib/' -i.orig gcc/config/i386/t-linux64;
-        #mkdir -pv "$source_packager_target"/libgcc
-        #ln -s "$SRCDIR"/libgcc/gthr-posix.h "$source_packager_target"/libgcc/gthr-default.h;
+        sh contrib/download_prerequisites;
+        sed -e '/m64=/s/lib64/lib/' -i.orig gcc/config/i386/t-linux64;
+        mkdir -pv "$source_packager_target"/libgcc
+        ln -s "$SRCDIR"/libgcc/gthr-posix.h "$source_packager_target"/libgcc/gthr-default.h;
     cd -;
 
     # Configure and build a barely useable gcc
     cd "$BLDDIR";
-    echo "$source_packager_builder_environment";
-        # "$SRCDIR"/configure \
-        #     --build=$("$SRCDIR"/config.guess) \
-        #     --enable-initfini-array \
-        #     --disable-bootstrap \
-        #     --disable-nls \
-        #     --disable-multilib \
-        #     --disable-decimal-float \
-        #     --disable-libatomic \
-        #     --disable-libgomp \
-        #     --disable-libquadmath \
-        #     --disable-libssp \
-        #     --disable-libvtv \
-        #     --disable-libstdcxx \
-        #     --enable-languages=c,c++;
-        # make -j$source_packager_number_of_jobs; 
-        # make -j$source_packager_number_of_jobs install;
-        # ln -sv gcc "$source_packager_builder_environment"/bin/cc;
+        "$SRCDIR"/configure \
+            --build=$("$SRCDIR"/config.guess) \
+            --prefix="$source_packager_builder_environment" \
+            --enable-initfini-array \
+            --enable-shared \
+            --disable-bootstrap \
+            --disable-multilib \
+            --enable-languages=c,c++;
+        make -j$source_packager_number_of_jobs; 
+        make -j$source_packager_number_of_jobs install;
+        ln -sv gcc "$source_packager_builder_environment"/bin/cc;
         cd -;
-}
-
-get_kernel_headers() {
-    LINUX_VERSION=$(uname -r | cut -c 1-3)
-    FILEURL="https://github.com/torvalds/linux/archive/v$LINUX_VERSION.tar.gz"
-    NAME="linux-$LINUX_VERSION";
-    SRCDIR="$source_packager_source_dir/$NAME"
-
-    if [ -d "$SRCDIR" ]; then
-        rm -rfv "$SRCDIR";
-    fi;
-
-    from_tz_file "$FILEURL" "$NAME";
-
-    cd "$SRCDIR"; 
-        make -j$source_packager_number_of_jobs mrproper && \
-        make -j$source_packager_number_of_jobs headers;
-        find usr/include -name '.*' -delete;
-        rm usr/include/Makefile;
-        cp -rfv usr/include/* "$source_packager_include"/;
-    cd -;
-    rm -rf "$BLDDIR";
 }
 
 glibc_stage_two() {
@@ -114,19 +86,14 @@ glibc_stage_two() {
         rm -rf "$BLDDIR";
     fi;
 
-    if ! [ -f "$source_packager_library"/ld-lsb.so.3 ]; then
-        ln -sv /lib/ld-linux.so.2 "$source_packager_library"/ld-lsb.so.3;
-        ln -sfv /usr/lib64/ld-linux-x86-64.so.2 "$source_packager_library_64";
-        ln -sfv "$source_packager_library_64"/ld-linux-x86-64.so.2 "$source_packager_library_64"/ld-lsb-x86-64.so.3;
-    fi;
-
     cd "$SRCDIR";
         mkdir "$BLDDIR" -pv;
         pushd "$BLDDIR";
             "$SRCDIR"/configure \
                 --prefix="$source_packager_builder_environment" \
-                --build=$("$SRCDIR"/scripts/config.guess) \
+                --disable-werror \
                 --enable-kernel=3.2 \
+                --enable-stack-protector=strong \
                 --with-headers="$source_packager_include" \
                 --without-selinux \
                 libc_cv_slibdir="$source_packager_library";
@@ -134,7 +101,6 @@ glibc_stage_two() {
                 make -j$source_packager_number_of_jobs install;
         popd;
     cd -;
-    "$source_packager_builder_environment"/libexec/gcc/"$source_packager_target"/10.2.0/install-tools/mkheaders
 }
 
 lib_cpp_stage_two() {
@@ -152,7 +118,6 @@ lib_cpp_stage_two() {
     mkdir "$BLDDIR" -pv;
     cd "$BLDDIR";
         "$SRCDIR"/libstdc++-v3/configure \
-            --build=$("$SRCDIR"/config.guess) \
             --prefix="$source_packager_builder_environment" \
             --disable-multilib \
             --disable-nls \
@@ -166,6 +131,5 @@ lib_cpp_stage_two() {
 
 binutils_stage_two;
 # gcc_stage_two;
-# get_kernel_headers;
 # glibc_stage_two;
 # lib_cpp_stage_two;
